requirements.txt
Flask==2.0.1
line-bot-sdk==1.20.0
opencv-python-headless==4.5.3.56
torch==1.9.0
